+++ tests/src/rules/extensions.js
@@ -63,7 +63,7 @@ ruleTester.run('extensions', rule, {
       code: `
         import foo from './foo.js'
         import bar from './bar.json'
-        import Component from './Component'
+        import Component from './Component.jsx'
         import express from 'express'
       `,
       options: [ 'ignorePackages' ],
@@ -296,6 +296,28 @@ ruleTester.run('extensions', rule, {
         import express from 'express'
       `,
       options: [ 'always', {ignorePackages: true} ],
+      errors: [
+        {
+          message: 'Missing file extension for "./Component"',
+          line: 4,
+          column: 31,
+        }, {
+          message: 'Missing file extension for "foo/baz"',
+          line: 5,
+          column: 25,
+        },
+      ],
+    }),
+
+    test({
+      code: `
+        import foo from './foo.js'
+        import bar from './bar.json'
+        import Component from './Component'
+        import baz from 'foo/baz'
+        import express from 'express'
+      `,
+      options: [ 'ignorePackages' ],
       errors: [
         {
           message: 'Missing file extension for "./Component"',